### 2019-11-21 #1907 @kthyng
Hi @michael-hartmann and thanks for your submission. Please add to your paper to [include](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain):

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

and possibly also

> A clear Statement of Need that illustrates the research purpose of the software

which is hard to discern from this outside perspective currently. We want to see a clear research use stated clearly.
