+++ docs/paper/paper.bib
@@ -170,3 +170,31 @@
   year    = {2006},
   doi     = {10.1103/PhysRevLett.96.080403}
 }
+
+% +-------------+
+% | Experiments |
+% +-------------+
+
+% https://doi.org/10.1103/PhysRevB.100.081406
+@article{liu_prb_2019,
+  author  = {Liu, M. and Xu, J. and Klimchitskaya, G. L. and Mostepanenko, V. M. and Mohideen, U.},
+  title   = {{Examining the Casimir puzzle with an upgraded AFM-based technique and advanced surface cleaning}},
+  journal = {Physical Review B},
+  volume  = {100},
+  issue   = {8},
+  pages   = {081406},
+  year    = {2019},
+  doi     = {10.1103/PhysRevB.100.081406}
+}
+
+% https://journals.aps.org/pra/abstract/10.1103/PhysRevA.100.052511
+@article{liu_pra_2019,
+  author  = {Liu, M. and Xu, J. and Klimchitskaya, G. L. and Mostepanenko, V. M. and Mohideen, U.},
+  title   = {{Precision measurements of the gradient of the Casimir force between ultraclean metallic surfaces at larger separations}},
+  journal = {Physical Review A},
+  volume  = {100},
+  issue   = {5},
+  pages   = {052511},
+  year    = {2019},
+  doi     = {10.1103/PhysRevA.100.052511}
+}
