+++ docs/paper/paper.md
@@ -25,16 +25,28 @@ bibliography: paper.bib
 # Summary
 
 CaPS is a package for the analysis of the Casimir effect in the plane-sphere
-geometry. This geometry depicted in Fig. 1 is most commonly used in precision
-measurements of the Casimir force arising from quantum and thermal fluctuations
-of the electromagnetic field [@bordag_advances]. Specifically, CaPS allows to
-compute the Casimir free energy as a function of the sphere radius $R$, the
-minimal separation $L$ between sphere and plane, the temperature $T$, and the
-material properties of plane and sphere. It is assumed that both objects are
+geometry. The Casimir force arises due to quantum and thermal fluctuations of
+the electromagnetic field and is closely related to the van der Waals forces
+[@bordag_advances]. As an observable consequence of vacuum fluctuations in the
+macroscopic world, the Casimir effect is linked to the problem of zero-point
+energy and the cosmological constant problem. The Casimir force is the dominant
+force between neutral non-magnetic materials in the nanometre to micrometre
+range. As such, a precise knowledge of the Casimir force is crucial in order to
+detect possible deviations from the gravitational interaction, and thus to
+either exclude or possibly support proposed mechanisms a fifth fundamental
+interaction. From a technological point of view, the Casimir effect has been
+shown to play an important role for micro- and nano-electromechanical systems.
+
+CaPS allows to compute the Casimir interaction in the plane-sphere geometry as
+shown in Fig. 1. The plane-sphere geometry is most commonly used in precision
+measurements of the Casimir force. Specifically, CaPS allows to compute the
+Casimir free energy as a function of the sphere radius $R$, the minimal
+separation $L$ between sphere and plane, the temperature $T$, and the material
+properties of plane and sphere. It is assumed that both objects are
 non-magnetic and placed in vacuum. In typical experiments the aspect ratio
-$R/L$ is of the order of 1000. The main purpose of this package is to
-make aspect ratios as large as $R/L\sim5000$ accessible. Higher aspect ratios
-are usually well covered by the proximity force approximation.
+$R/L$ is of the order of 1000. The main purpose of this package is to make
+aspect ratios as large as $R/L\sim5000$ accessible. Higher aspect ratios are
+usually well covered by the proximity force approximation.
 
 ![Geometry of the plane-sphere setup: A sphere with radius $R$ is separated by
 the distance $L$ from an infinitely extended plate. In typical experiments,
@@ -87,7 +99,7 @@ The implementation for the plane-cylinder geometry is based on [@emig_prl_2006].
 
 ``CaPS`` has been used to analyze negative Casimir entropies [@ingold_pre_2015;
 @umrath_pre_2015], and to study corrections to the widely used proximity force
-approximation for experimentally relevant parameters [@hartmann_prl_2017].  In
+approximation for experimentally relevant parameters [@hartmann_prl_2017]. In
 addition, data generated using ``CaPS`` was used to analyze the experiments in
 Ref. [@liu_prb_2019] and Ref. [@liu_pra_2019]. The package is released under
 the GPLv2 license.
