### 2019-12-15 5f33fc8
Mention that using the free energy the force can be computed
