### 2019-12-15 174b79d
Rephrasing / changing the emphasis
