### 2019-12-15 8eb8445
Add (R) for rapid communication to Liu, PRB reference
