+++ docs/paper/paper.bib
@@ -182,7 +182,7 @@
   journal = {Physical Review B},
   volume  = {100},
   issue   = {8},
-  pages   = {081406},
+  pages   = {081406(R)},
   year    = {2019},
   doi     = {10.1103/PhysRevB.100.081406}
 }
