+++ docs/paper/paper.bib
@@ -1,11 +1,64 @@
+% +--------------+
+% | Introduction |
+% +--------------+
+
+% https://doi.org/10.1016/j.crhy.2011.05.004
+@article{antoniadis_crp_2011,
+  title   = {Short-range fundamental forces},
+  author  = {I. Antoniadis and S. Baessler and M. B\"uchner and V.V. Fedorov and S. Hoedl and A. Lambrecht and V.V. Nesvizhevsky and G. Pignol and K.V. Protasov and S. Reynaud and Yu. Sobolev},
+  author  = {Antoniadis, I. and Baessler, S. and B\"uchner, M. and Fedorov, V. V. and Hoedl, S. and Lambrecht, A. and Nesvizhevsky, V. V. and Pignol, G. and Protasov, K. V. and Reynaud, S. and Sobolev, Yu.},
+  journal = {C. R. Phys.},
+  volume  = {12},
+  number  = {8},
+  pages   = {755--778},
+  year    = {2011},
+  doi     = {10.1016/j.crhy.2011.05.004}
+}
+
+% https://doi.org/10.1103/PhysRevLett.87.211801
+@article{chan_prl_2001,
+  title   = {Nonlinear Micromechanical Casimir Oscillator},
+  author  = {Chan, H. B. and Aksyuk, V. A. and Kleiman, R. N. and Bishop, D. J. and Capasso, Federico},
+  journal = {Phys. Rev. Lett.},
+  volume  = {87},
+  issue   = {21},
+  pages   = {211801},
+  year    = {2001},
+  doi     = {10.1103/PhysRevLett.87.211801}
+}
+
+% https://doi.org/10.1103/PhysRevB.63.033402
+@article{buks_prb_2001,
+  title   = {Stiction, adhesion energy, and the Casimir effect in micromechanical systems},
+  author  = {Buks, E. and Roukes, M. L.},
+  journal = {Phys. Rev. B},
+  volume  = {63},
+  issue   = {3},
+  pages   = {033402},
+  year    = {2001},
+  doi     = {10.1103/PhysRevB.63.033402}
+}
+
 % https://doi.org/10.1093/acprof:oso/9780199238743.001.0001
 @book{bordag_advances,
   title     = {{Advances in the Casimir Effect}},
-  author    = {Bordag, M. and Klimchitskaya, G. L. and Mohideen, U. and Mostepanenko, V. M.}, 
+  author    = {Bordag, M. and Klimchitskaya, G. L. and Mohideen, U. and Mostepanenko, V. M.},
   year      = {2009},
   publisher = {Oxford University Press},
   isbn      = {978-0-199238-743},
   doi       = {10.1093/acprof:oso/9780199238743.001.0001}
+}
+
+% https://doi.org/10.1016/j.crhy.2012.04.008
+@article{martin_crp_2012,
+  title   = {Everything you always wanted to know about the cosmological constant problem (but were afraid to ask)},
+  author  = {Martin, J.},
+  journal = {C. R. Phys.},
+  volume  = {13},
+  number  = {6},
+  pages   = {566--665},
+  year    = {2012},
+  doi     = {10.1016/j.crhy.2012.04.008}
 }
 
 @phdthesis{Durand_phd,
