+++ docs/paper/paper.md
@@ -31,11 +31,12 @@ the electromagnetic field and is closely related to the van der Waals force
 materials in the nanometre to micrometre range and plays an important role in
 colloidal systems. Of technological relevance are applications to micro- and
 nano-electromechanical systems where the Casimir force can lead to stiction and
-thus constitute a failure mechanism. On a more fundamental level, the Casimir
-effect is linked to the zero-point energy and the cosmological constant
-problem. A precise knowledge of the Casimir force is crucial for the search for
-possible deviations from Newton's law of gravitation which could arise from a
-fifth fundamental interaction.
+thus constitute a failure mechanism [@chan_prl_2001; @buks_prb_2001]. On a more
+fundamental level, the Casimir effect is linked to the zero-point energy and
+the cosmological constant problem [@martin_crp_2012]. A precise knowledge of
+the Casimir force is crucial for the search for possible deviations from
+Newton's law of gravitation which could arise from a fifth fundamental
+interaction [@antoniadis_crp_2011].
 
 CaPS allows to compute the Casimir interaction in the plane-sphere geometry as
 shown in Fig. 1. The plane-sphere geometry is most commonly used in precision
