### 2019-12-16 #1907 @michael-hartmann
@kthyng, @danielskatz

Thank you for your suggestions.

We have addressed the previous points by (i) extending the introduction for a non-specific audience, and (ii) adding references to experiments that have used data generated with our software.

We hope that the review can now continue. :-)
