### 2019-12-30 d8de5e4
Declare functions static to avoid compiler warning
