### 2019-12-30 c79cf4e
Also recognize MSVC compiler
