### 2019-12-30 57c9a75
Improve CMakeLists.txt file
* Check if linking against libm is needed
* Set warnings correctly if compiler is MSVC
* Use cmake function instead of uname to determine info about host
* Use cmake function to find git executable
* Correctly specify link targets hodlr and caps
