+++ CMakeLists.txt
@@ -1,25 +1,37 @@
-cmake_minimum_required (VERSION 3.0)
+cmake_minimum_required (VERSION 3.6)
 project (CaPS)
 
+# git is optional
+find_package(Git)
+
+# required packages: mpi, blas, and lapack
 find_package(MPI REQUIRED)
 find_package(BLAS REQUIRED)
 find_package(LAPACK REQUIRED)
+
+# check if we need to link against libm
+check_library_exists(m sqrt "" LIBM)
 
 include_directories(${MPI_INCLUDE_PATH})
 include_directories("src/include/")
 include_directories("src/cquadpack/include/")
 include_directories("src/libhodlr/include/")
 
-# set standard optimization -O3
-if(NOT DEFINED OPT)
-    set(OPT "-O3")
+if(NOT CMAKE_BUILD_TYPE)
+  set(CMAKE_BUILD_TYPE Release)
 endif()
 
-# set CFLAGS
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -std=c99 ${OPT} -Wall -Wextra -Wmissing-prototypes -Wshadow -Wpointer-arith -Wcast-qual -Wwrite-strings -Wno-unused-parameter -fstrict-aliasing")
+set(CMAKE_C_FLAGS_DEBUG "-g")
+set(CMAKE_C_FLAGS_RELEASE "-O3")
 
-# set CXXFLAGS
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 ${OPT} -Wall")
+# set warnings
+if(MSVC)
+    set(CMAKE_C_FLAGS "/W2")
+    set(CMAKE_CXX_FLAGS "/W2")
+else()
+    set(CMAKE_C_FLAGS "-std=c99 -Wall -Wextra -Wmissing-prototypes -Wshadow -Wpointer-arith -Wcast-qual -Wwrite-strings -Wno-unused-parameter -fstrict-aliasing")
+    set(CMAKE_CXX_FLAGS "-std=c++11 -Wall")
+endif()
 
 # By default, icc violates strict IEEE floating point behaviour - similar to
 # GCC's --fast-math option. The code, however, relies on strict IEEE floating
@@ -43,76 +55,87 @@ add_library(caps SHARED src/bessel.c src/cache.c src/fcqs.c src/integration.c sr
 
 target_compile_definitions(caps PRIVATE LIBCAPS_VERSION="0.5")
 
-# get machine name
-execute_process(
-  COMMAND uname -smnr
-  WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
-  OUTPUT_VARIABLE MACHINE
-  OUTPUT_STRIP_TRAILING_WHITESPACE
-)
-target_compile_definitions(caps PRIVATE -DMACHINE="${MACHINE}")
+# get information about host
+cmake_host_system_information(RESULT hostname QUERY HOSTNAME)
+cmake_host_system_information(RESULT processor_description QUERY PROCESSOR_DESCRIPTION)
+cmake_host_system_information(RESULT os_name QUERY OS_NAME)
+cmake_host_system_information(RESULT os_release QUERY OS_RELEASE)
+cmake_host_system_information(RESULT os_version QUERY OS_VERSION)
+cmake_host_system_information(RESULT os_platform QUERY OS_PLATFORM)
+set(host_info_raw "${hostname}, ${processor_description}, ${os_name} ${os_version} (${os_platform})")
+string(REPLACE "#" "" host_info ${host_info_raw})
+target_compile_definitions(caps PRIVATE -DMACHINE="${host_info}")
 
[…]
