### 2019-12-30 7da90d2
Change package for libc, add remark about clang
* Instead of libc6-dev give libc-dev as requirement on Debian/Ubuntu
* Mention that compiling the tests with clang takes a couple of minutes
