### 2019-12-30 f563128
Correctly link libcaps, set optimization correctly for msvc
