### 2019-12-30 d9e5e35
Fix link targets, don't pass -Wno-unknown-pragmas to MSVc
