### 2019-12-30 ad63365
Add function to compare strings case-insensitive
