### 2019-12-30 a746f52
Avoid using strings.h and strcasecmp
