### 2019-12-30 d234b54
Avoid variable length arrays (VLA)
