### 2019-12-30 21c978a
Declare used LAPACK functions in code instead of including clapack.h
