### 2019-12-30 4b10314
Declare used LAPACK functions in code instead of including clapack.h
