### 2019-12-30 fb542b4
Avoid array arguments in function psd
