### 2019-12-30 1ce45eb
Avoid array arguments in function psd
