### 2019-12-30 4bd9970
Avoid variable length arrays (VLAs)
