### 2019-12-30 f267991
Avoid variable length array (VLA)
