+++ src/libcaps.c
@@ -10,7 +10,6 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <string.h>
-#include <unistd.h>
 
 #include "quadpack.h"
 #include "constants.h"
