### 2019-12-30 5a55b6f
Avoid UNIX specific functions
* Remove line printing PID
* Use workaround for usleep on Windows
* Do not use access to test if file is readable
