### 2019-12-31 cc47186
Avoid header file
