### 2019-12-31 83662d1
Remove time_as_string, add strtodouble
