### 2019-12-31 56702b0
Add time_as_string, use strtodouble instead of atof
