### 2019-12-31 f8c9c38
Use strtodouble instead of atof
