### 2019-12-31 2980d09
Use strtodouble instead of atof, don't modify environment
