### 2019-12-31 75bd647
Define __attribute__ for compilers that do not support it
