+++ src/include/logfac.h
@@ -6,6 +6,8 @@
 extern "C" {
 #endif
 
+#include "attributes.h"
+
 double lfac(unsigned int n) __attribute__ ((pure));
 double logi(unsigned int x) __attribute__ ((pure));
 
