### 2019-12-31 6bbc6c6
Remove declaration of time_as_string
