+++ src/include/utils.h
@@ -42,7 +42,6 @@ extern "C" {
 #define xfree(p) do { free(p); p = NULL; } while(0)
 
 double now(void);
-void time_as_string(char *s, size_t len);
 
 void *xmalloc(size_t size);
 void *xrealloc(void *p, size_t size);
