+++ src/include/misc.h
@@ -5,11 +5,8 @@
 extern "C" {
 #endif
 
-#include <stdlib.h>
-#include <math.h>
-
 #include "attributes.h"
-#include "constants.h"
+#include "libcaps.h"
 
 /** represent number \f$v\f$ by its sign and \f$\log|v|\f$ */
 typedef struct
