### 2019-12-31 38c7172
Add constants previously in constants.h
