+++ src/include/libcaps.h
@@ -23,7 +23,6 @@ typedef enum { TE, TM } polarization_t;
 #include "attributes.h"
 #include "cache.h"
 #include "matrix.h"
-#include "constants.h"
 
 
 /* default values */
@@ -32,6 +31,36 @@ typedef enum { TE, TM } polarization_t;
 #define CAPS_EPSREL 1e-8  /**< default relative error for integration */
 
 #define CAPS_CACHE_ELEMS 10000000 /**< default number of elems of the cache for I integrals */
+
+/*! macro to get minimum of two numbers */
+#ifndef MIN
+#define MIN(a,b) (((a)<(b))?(a):(b))
+#endif
+
+/*! macro to get maximum of two numbers */
+#ifndef MAX
+#define MAX(a,b) ((((a))>((b)))?((a)):((b)))
+#endif
+
+/*! macro to get sign of numbers */
+#define SGN(val) ((0 < (val)) - ((val) < 0))
+
+/*! compute \f$x^2\f$ */
+#define POW_2(x) ((x)*(x))
+
+#define CAPS_PI      3.14159265358979323846 /**< value for \f$\pi=3.141592...\f$ */
+#define CAPS_LOG2    0.6931471805599453 /**< \f$\log(2)\f$ */
+#define CAPS_LOGPI   1.1447298858494002 /**< \f$\log(\pi)\f$ */
+#define CAPS_HBAR    1.0545718e-34   /**< reduced Planck constant \f$\hbar\f$ [m² kg / s] */
+#define CAPS_HBAR_EV 6.582119514e-16 /**< reduced Planck constant \f$\hbar\f$ [eV s/rad] */
+#define CAPS_KB      1.38064852e-23  /**< Boltzman constant \f$k_\mathrm{B}\f$ [m² kg / ( K s² )] */
+#define CAPS_C       299792458.      /**< speed of light \f$c\f$ in vacuum [m/s] */
+
+/**
+ * define sign_t as a signed char, because "char can be either signed or
+ * unsigned depending on the implementation"
+ */
+typedef signed char sign_t;
 
 /**
  * The CaPS object. This structure stores all essential information about
