### 2019-12-31 3e3caab
Don't include constants, adjust name of constants
