### 2019-12-31 be9a910
Do not include constants.h, adjust name of constants
