### 2019-12-31 e6d10b1
Make now() work on win32
