### 2019-12-31 b4f5a6f
Use CAPS_PI or define M_PI
