+++ src/integration.c
@@ -134,7 +134,7 @@ static double K_estimate(int nu, int m, double alpha, double eps, double *a, dou
             *a = 1;
             *b = 1-log(eps)/alpha;
 
-            const double logt1 = 1.5*log(alpha)+log(2/M_PI)/2+bessel_logKn_half(nu,alpha);
+            const double logt1 = 1.5*log(alpha)+log(2/CAPS_PI)/2+bessel_logKn_half(nu,alpha);
             const double logt2 = -alpha+log(alpha+nu*(nu+1)/2.);
             const double arg = -exp(logt2-logt1);
 
@@ -199,7 +199,7 @@ static double K_estimate(int nu, int m, double alpha, double eps, double *a, dou
 
     {
         /* approximation using Laplace's method */
-        *approx = log(2*M_PI/fpp)/2-fxmax;
+        *approx = log(2*CAPS_PI/fpp)/2-fxmax;
 
         /* estimate width, left and right borders */
         const double width = -log(eps)/sqrt(fpp);
@@ -308,7 +308,7 @@ static double _caps_integrate_K(integration_t *self, int nu, polarization_t p, s
              * integral: 2*exp(-α) + sqrt(2/(pi*α)) [ (ν+1)(ν+2)K_{ν+1/2}(α) - 2α K_(ν+3/2)(α) ]
              *           \- t0 --/   \----- C ----/   \------- t1 ---------/   \---- t2 -----/
              */
-            const double logC = log(2/(alpha*M_PI))/2;
+            const double logC = log(2/(alpha*CAPS_PI))/2;
 
             log_t terms[3];
 
@@ -328,9 +328,9 @@ static double _caps_integrate_K(integration_t *self, int nu, polarization_t p, s
             return logadd_ms(terms, 3, &dummy);
         }
         else if(nu == (2*m))
-            return -0.5*log(M_PI)+(m-0.5)*log(2/alpha)+lfac(m-1)+lfac2(4*m-1)+bessel_logKnu_half(m-1,alpha);
+            return -0.5*log(CAPS_PI)+(m-0.5)*log(2/alpha)+lfac(m-1)+lfac2(4*m-1)+bessel_logKnu_half(m-1,alpha);
         else if(nu == (2*m+1))
-            return -0.5*log(M_PI)+(m-0.5)*log(2/alpha)+lfac(m-1)+lfac2(4*m-1)+bessel_logKnu_half(m,alpha)+logi(4*m+1);
+            return -0.5*log(CAPS_PI)+(m-0.5)*log(2/alpha)+lfac(m-1)+lfac2(4*m-1)+bessel_logKnu_half(m,alpha)+logi(4*m+1);
         else if(m == 1)
         {
             /* use analytical result for m=1 and PR
@@ -352,7 +352,7 @@ static double _caps_integrate_K(integration_t *self, int nu, polarization_t p, s
                 /* integral: α*exp(-α)*( ae1 + ae2 + ae3 + ... -ν(ν+1)/(2α)) */
                 return log(alpha)-alpha+log(ae1+ae2+ae3-nu*(nu+1.)/(2*alpha));
 
-            const double logt1 = 1.5*log(alpha)+log(2/M_PI)/2+bessel_logKnu_half(nu,alpha);
+            const double logt1 = 1.5*log(alpha)+log(2/CAPS_PI)/2+bessel_logKnu_half(nu,alpha);
             const double logt2 = -alpha+log(alpha+nu*(nu+1.)/2.);
 
             return logt1 + log1p(-exp(logt2-logt1));
