+++ src/bessel.c
@@ -11,6 +11,11 @@
 
 #include "bessel.h"
 
+#ifndef M_PI
+#define M_PI 3.14159265358979323846
+#endif
+
+
 /**
  * @name modified Bessel functions for orders \f$n=0,1\f$
  * @{
