### 2020-01-01 db8f406
Make all functions static, avoid header file
