+++ src/caps.c
@@ -16,8 +16,6 @@
 #include <unistd.h>
 #endif
 
-#include "caps.h"
-
 #include "buf.h"
 #include "fcqs.h"
 #include "libcaps.h"
@@ -27,6 +25,7 @@
 #include "quadpack.h"
 #include "utils.h"
 
+/* preprocessor define */
 #define EPSREL 1e-6 /**< default value for --epsrel */
 #define CUTOFF 1e-9 /**< default value for --cutoff */
 #define LDIM_MIN 20 /**< minimum value for --ldim */
@@ -35,6 +34,43 @@
 
 #define STATE_RUNNING 1
 #define STATE_IDLE    0
+
+/* local typedefs */
+
+typedef struct {
+    int index, m;
+    double xi_;
+    double recv;
+    double value;
+    MPI_Request request;
+    int state;
+} caps_task_t;
+
+typedef struct {
+    double L, R, T, omegap, gamma, cutoff, iepsrel, alpha;
+    int ldim, cores;
+    bool verbose;
+    caps_task_t **tasks;
+    int determinants;
+    char filename[512];
+    double cache[4096][2];
+    int cache_elems;
+} caps_mpi_t;
+
+/* prototypes */
+static caps_mpi_t *caps_mpi_init(double L, double R, double T, char *filename, char *resume, double omegap, double gamma_, int ldim, double cutoff, double iepsrel, int cores, bool verbose);
+static void caps_mpi_free(caps_mpi_t *self);
+static int caps_mpi_submit(caps_mpi_t *self, int index, double xi, int m);
+static int caps_mpi_retrieve(caps_mpi_t *self, caps_task_t **task_out);
+static int caps_mpi_get_running(caps_mpi_t *self);
+static int caps_get_determinants(caps_mpi_t *self);
+
+static void usage(FILE *stream);
+
+static void F_HT(caps_mpi_t *caps_mpi, double omegap, double *drude, double *pr, double *plasma);
+static double F_xi(double xi, caps_mpi_t *caps_mpi);
+static void master(int argc, char *argv[], const int cores);
+static void slave(MPI_Comm master_comm, const int rank);
 
 
 /** @brief Write time into string
@@ -92,7 +128,7 @@ static bool is_readable(const char *filename)
  * @param [in] verbose flag if verbose
  * @retval object caps_mpi_t object
  */
-caps_mpi_t *caps_mpi_init(double L, double R, double T, char *filename, char *resume, double omegap, double gamma_, int ldim, double cutoff, double iepsrel, int cores, bool verbose)
+static caps_mpi_t *caps_mpi_init(double L, double R, double T, char *filename, char *resume, double omegap, double gamma_, int ldim, double cutoff, double iepsrel, int cores, bool verbose)
 {
     caps_mpi_t *self = xmalloc(sizeof(caps_mpi_t));
 
@@ -178,7 +214,7 @@ static void _mpi_stop(int cores)
  *
  * @param [in] self caps_mpit_t object
  */
-void caps_mpi_free(caps_mpi_t *self)
+static void caps_mpi_free(caps_mpi_t *self)
 {
     _mpi_stop(self->cores);
 
@@ -195,7 +231,7 @@ void caps_mpi_free(caps_mpi_t *self)
  * @param [in] self caps_mpit_t object
  * @retval running number of processes that are running
  */
-int caps_mpi_get_running(caps_mpi_t *self)
+static int caps_mpi_get_running(caps_mpi_t *self)
 {
     int running = 0;
 
[…]
