### 2020-01-01 54625f9
Do not use __attribute__ for static functions
