### 2020-01-04 44badb1
Avoid preprocessor symbols starting with __
