### 2020-01-04 59b4ec2
Avoid preprocessor symbols starting with __
