### 2020-01-04 7a55a6d
Fix names of constants
