### 2020-01-04 b6c1aa9
Remove make from commands because it's redundant
