### 2020-01-04 c15d60b
Improve CMakeLists.txt file
* build code position independent
* correctly specify dependencies of targets
* add target for cquadpack and argparse
* by default, build static libraries
* add option to build libcaps as a shared library
* change optimization vor MSVC to /O2 as /O3 doesn't exist
* remove some not so useful output messages
