+++ CMakeLists.txt
@@ -1,6 +1,8 @@
 cmake_minimum_required (VERSION 3.6)
 project (CaPS)
 
+option(BUILD_SHARED "Build libcaps as shared library" OFF)
+
 # git is optional
 find_package(Git)
 
@@ -9,24 +11,33 @@ find_package(MPI REQUIRED)
 find_package(BLAS REQUIRED)
 find_package(LAPACK REQUIRED)
 
+# generate position independent code for all targets
+set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+
 # check if we need to link against libm
 check_library_exists(m sqrt "" LIBM)
 
+# add necessary include directories
 include_directories(${MPI_INCLUDE_PATH})
 include_directories("src/include/")
 include_directories("src/cquadpack/include/")
 include_directories("src/libhodlr/include/")
 
+# set release to default build type
 if(NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE Release)
 endif()
 
 if(MSVC)
-    set(CMAKE_C_FLAGS_RELEASE "/O3")
+    set(CMAKE_C_FLAGS_RELEASE "/O2")
+    set(CMAKE_CXX_FLAGS_RELEASE "/O2")
     set(CMAKE_C_FLAGS_DEBUG "/g")
+    set(CMAKE_CXX_FLAGS_DEBUG "/g")
 else()
     set(CMAKE_C_FLAGS_RELEASE "-O3")
+    set(CMAKE_CXX_FLAGS_RELEASE "-O3")
     set(CMAKE_C_FLAGS_DEBUG "-g")
+    set(CMAKE_CXX_FLAGS_DEBUG "-g")
 endif()
 
 # set warnings
@@ -51,14 +62,25 @@ if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Intel")
 endif()
 
 # libhodlr
-add_library(hodlr SHARED src/libhodlr/src/hodlr.cpp src/libhodlr/src/HODLR_Matrix.cpp src/libhodlr/src/HODLR_Node.cpp src/libhodlr/src/HODLR_Tree.cpp src/libhodlr/src/HODLR_Tree_NonSPD.cpp src/libhodlr/src/HODLR_Tree_SPD.cpp src/libhodlr/src/KDTree.cpp)
+add_library(hodlr STATIC src/libhodlr/src/hodlr.cpp src/libhodlr/src/HODLR_Matrix.cpp src/libhodlr/src/HODLR_Node.cpp src/libhodlr/src/HODLR_Tree.cpp src/libhodlr/src/HODLR_Tree_NonSPD.cpp src/libhodlr/src/HODLR_Tree_SPD.cpp src/libhodlr/src/KDTree.cpp)
 target_compile_definitions(hodlr PRIVATE -DUSE_DOUBLE)
 if(NOT MSVC)
     set_target_properties(hodlr PROPERTIES COMPILE_FLAGS "-Wno-unknown-pragmas")
 endif()
 
+# cquadpack
+add_library(cquadpack STATIC src/cquadpack/src/dqage.c src/cquadpack/src/dqagi.c src/cquadpack/src/dqags.c src/cquadpack/src/dqext.c src/cquadpack/src/dqk15.c src/cquadpack/src/dqk15i.c src/cquadpack/src/dqk21.c src/cquadpack/src/dqk31.c src/cquadpack/src/dqk41.c src/cquadpack/src/dqk51.c src/cquadpack/src/dqk61.c src/cquadpack/src/dqsort.c)
+
+# argparse
+add_library(argparse STATIC src/argparse.c)
+
 # libcaps
[…]
