### 2020-01-06 f61e83a
Add functions caps_detalg_from_string, caps_detalg_to_string
