### 2020-01-06 e69b9d2
Add function argparse_usage_stream
