### 2020-01-06 bf8868f
Replace getopt by argparse
