+++ src/caps_logdetD.c
@@ -1,46 +1,11 @@
-#include <getopt.h>
-#include <math.h>
 #include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 
+#include "argparse.h"
 #include "material.h"
 #include "libcaps.h"
-#include "misc.h"
-#include "utils.h"
 
-
-/* print usage */
-static void usage(FILE *stream)
-{
-    fprintf(stream,
-"Usage: capc_logdetD [OPTIONS]\n\n"
-"This program will calculate the free Casimir energy for the plane-sphere\n"
-"geometry for given n,m,T,L/R.\n"
-"\n"
-"Mandatory options:\n"
-"    -L     separation L\n"
-"    -R     radius R\n"
-"    --xi   imaginary frequency ξ in units of (L+R)/c\n"
-"    -m     value of m\n"
-"\n"
-"Further options:\n"
-"    -l, --ldim LDIM\n"
-"        Set ldim to LDIM.\n"
-"\n"
-"    -f, --material FILENAME\n"
-"        Use material described by FILENAME.\n"
-"\n"
-"    -d, --detalg DETALG\n"
-"        Compute the matrix using DETALG (LU, QR, CHOLESKY, HODLR)\n"
-"\n"
-"    -i, --iepsrel IEPSREL\n"
-"        Relative accuracy to evaluate integrals\n"
-"\n"
-"    -h,--help\n"
-"        Show this help.\n"
-"\n"
-"\n"
+static const char *usage_epilog =
+"\n\n"
 "Environment variables:\n"
 "   CAPS_DUMP:\n"
 "        If this variable is set, the round-trip matrix will be dumped in numpy\n"
@@ -48,100 +13,47 @@ static void usage(FILE *stream)
 "        round-trip matrix will only be dumped if detalg is QR, LU or CHOLESKY.\n"
 "\n"
 "   CAPS_CACHE_ELEMS:\n"
-"        Determines the size of the cache for the integrals I."
-"\n");
-}
+"        Determines the size of the cache for the integrals I.";
 
-int main(int argc, char *argv[])
+int main(int argc, const char **argv)
 {
-    double iepsrel = 0;
     double start_time = now();
-    detalg_t detalg = DETALG_HODLR;
 
-    char filename[512] = { 0 };
-
-    /* geometry, Matsubara frequency */
+    /* geometry, Matsubara frequency, filename */
     double L = 0, R = 0, xi_ = -1;
     int m = -1;
+    char *filename = NULL;
 
     /* numerical parameters */
     int ldim = 0;
+    double iepsrel = 0;
+    detalg_t detalg = DETALG_HODLR;
+    char *detalg_str = NULL;
 
-    while(1)
-    {
-        struct option long_options[] = {
-            { "help",      no_argument,       0, 'h' },
+    const char *const usage[] = {
+        "caps_logetD -L separation -R radius --xi frequency -m M [further options]",
+        NULL,
+    };
 
-            { "iepsrel",   required_argument, 0, 'i' },
-            { "detalg",    required_argument, 0, 'd' },
-            { "material",  required_argument, 0, 'f' },
-            { "xi",        required_argument, 0, 'x' },
-            { "ldim",      required_argument, 0, 'l' },
+    struct argparse_option options[] = {
+        OPT_HELP(),
+        OPT_GROUP("Mandatory options"),
+        OPT_DOUBLE('L', "separation", &L, "smallest seperation between plane and spheres in m", NULL, 0, 0),
+        OPT_DOUBLE('R', "radius", &R, "radius of the sphere in m", NULL, 0, 0),
[…]
