### 2020-01-06 f9dce65
Link caps_logdetD against argparse
