### 2020-01-07 #1907 @danielskatz
Sorry for the delay in getting back to this - our rotating associate editor-in-chief system showed some flaws over the holidays
