### 2020-01-07 #1907 @danielskatz
@michael-hartmann - do you have any suggestions for potential reviewers? If so, please mention them here in this thread, without @-tagging them. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission. Please take a look and see if you see any potentially-applicable reviewers.
