### 2020-01-07 4f0f8ce
Remove const from argv in prototype of main
