+++ src/caps.c
@@ -1,6 +1,5 @@
 #define _DEFAULT_SOURCE /* make usleep work */
 
-#include <getopt.h>
 #include <mpi.h>
 #include <stdbool.h>
 #include <stdio.h>
@@ -16,6 +15,7 @@
 #include <unistd.h>
 #endif
 
+#include "argparse.h"
 #include "buf.h"
 #include "fcqs.h"
 #include "libcaps.h"
@@ -65,13 +65,30 @@ static int caps_mpi_retrieve(caps_mpi_t *self, caps_task_t **task_out);
 static int caps_mpi_get_running(caps_mpi_t *self);
 static int caps_get_determinants(caps_mpi_t *self);
 
-static void usage(FILE *stream);
-
 static void F_HT(caps_mpi_t *caps_mpi, double omegap, double *drude, double *pr, double *plasma);
 static double F_xi(double xi, caps_mpi_t *caps_mpi);
-static void master(int argc, char *argv[], const int cores);
+static void master(int argc, char **argv, const int cores);
 static void slave(MPI_Comm master_comm, const int rank);
 
+static const char *usage_epilog =
+"\n"
+"This program computes the free Casimir energy E(T,L,R) for the plane-sphere\n"
+"geometry. L denotes the smallest separation between sphere and plane, R is the\n"
+"radius of the sphere, and T is the temperature.\n"
+"\n"
+"This program uses MPI for parallization and needs at least two cores to run.\n"
+"\n"
+"The free energy at T=0 is calculated using integration:\n"
+"   E(L,R,T=0) = ∫dξ log det(1-M(ξ)),  ξ=0...∞,\n"
+"where M denotes the round-trip operator. The integration is performed either\n"
+"using an adaptive Gauß-Kronrod quadrature or a Fourier-Chebshev quadrature\n"
+"scheme.\n"
+"\n"
+"References:\n"
+"  [1] Hartmann, The Casimir effect in the plane-sphere geometry and the\n"
+"      Proximity Force Approximation, phd thesis, 2018\n"
+"  [2] Hartmann, Ingold, Maia Neto, Advancing numerics for the Casimir effect\n"
+"      to experimentally relevant aspect ratios, Phys. Scr. 93, 114003 (2018)";
 
 /** @brief Write time into string
  *
@@ -148,9 +165,12 @@ static caps_mpi_t *caps_mpi_init(double L, double R, double T, char *filename, c
     /* number of determinants we have computed */
     self->determinants = 0;
 
-    TERMINATE(strlen(filename) > 511, "filename too long: %s", filename);
     memset(self->filename, '\0', sizeof(self->filename));
-    strncpy(self->filename, filename, sizeof(self->filename)-sizeof(char));
+    if(filename != NULL)
+    {
+        TERMINATE(strlen(filename) > 511, "filename too long: %s", filename);
+        strncpy(self->filename, filename, sizeof(self->filename)-sizeof(char));
+    }
 
     /* cache for resume */
     self->cache_elems = 0;
@@ -499,11 +519,10 @@ int main(int argc, char *argv[])
     return 0;
 }
 
-static void master(int argc, char *argv[], const int cores)
+static void master(int argc, char **argv, const int cores)
 {
-    bool verbose = false, fcqs = false, ht = false;
-    char filename[512] = { 0 };
-	char resume[512] = { 0 };
+    bool verbose = false, fcqs = false, ht = false, psd_ = false, version = false;
+    char *filename = NULL, *resume = NULL;
     int ldim = 0;
     double L = 0, R = 0, T = 0, omegap = INFINITY, gamma_ = 0;
     double cutoff = CUTOFF, epsrel = EPSREL, eta = ETA;
[…]
