### 2020-01-07 911093b
Adjust manual and README to the changes of the build script
