### 2020-01-07 #1907 @michael-hartmann
@danielskatz Thank you for your reply.

I had a look at the list of JOSS reviewers and think jochym and jwuttke are good candidates. Both have a background in theoretical physics, experience in scattering (might be helpful), and listed C among their programming languages.
