### 2020-01-07 #1907 @danielskatz
👋 @jochym and @jwuffke - Would either (or ideally) both of you be willing to review this for JOSS?
