### 2020-01-07 #1907 @danielskatz
👋 @jwuttke - Would you be willing to review this for JOSS?
