### 2020-01-09 #1907 @danielskatz
Sorry, I thought you were familiar with JOSS - see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for general info about the reviewing process, and https://joss.readthedocs.io/en/latest/review_criteria.html for the specific review criteria.  In brief, a JOSS review is of the short paper and the software repository, and it is open and collaborative, seeking to improve the paper and software to the point where they can be accepted.
