### 2020-01-13 #2011 @whedon
[REVIEW]: CaPS: Casimir Effect in the Plane-Sphere Geometry
**Submitting author:** @michael-hartmann (<a href="http://orcid.org/0000-0002-9245-8582">Michael Hartmann</a>)
**Repository:** <a href="https://github.com/michael-hartmann/caps/" target ="_blank">https://github.com/michael-hartmann/caps/</a>
**Version:** 0.5
