### 2020-01-13 #2011 @danielskatz
HI @eschnett & @jwuttke - thanks for agreeing to review this submission!

Your job is to review the [paper](https://github.com/openjournals/joss-papers/blob/joss.02011/joss.02011/10.21105.joss.02011.pdf) and the repository according to the criteria in your review list in the first comment in this issue, and check things off as you find them satisfactory. If you see things that you think need to be changed, open an issue in the source repository (and add a note about this review issue to link them to here), or if they are small, just state them here in this issue.

Please also read the first 2 comments in this issue carefully.

Let me know if you have any questions.
