### 2020-01-13 michael-hartmann/caps#150 @jwuttke
README.md: add instruction how to install & uninstall
as far as I see, there are only instructions on compiling, not on installing the software
