### 2020-01-13 #2011 @jwuttke
I submitted 7 issues. Otherwise I am very satisfied with project & paper, and I recommend acceptance in JOSS.
