### 2020-01-13 #2011 @danielskatz
Thanks!

Can you put the URL of this issue (https://github.com/openjournals/joss-reviews/issues/2011) in the issues you opened?  This helps us view them from here, where we can see if they are open or closed, for example
