### 2020-01-13 michael-hartmann/caps#150 @jwuttke
<- https://github.com/openjournals/joss-reviews/issues/2011
