### 2020-01-13 8528ff9
Use build/ as build directory instead of bin/
