### 2020-01-13 39a0396
Initialize variables
Initialize all variables to avoid compiler warnings.
