### 2020-01-15 #2011 @jwuttke
@michael-hartmann - My issues are suggestions, not imperatives. I hope you will find some of them useful, but you are free to disagree.
