### 2020-01-15 #2011 @michael-hartmann
@danielskatz I am working on the issues right now.

@jwuttke Thank you for your suggestions. I agree with your issues and I will fix them as soon as possible. :-)
