### 2020-01-15 1d47ac0
Build static library if BUILD_SHARED is false
