### 2020-01-15 329dc5c
Add comment that files were generated by script
Add a comment to logfac.c and logfac.h that the files were generated by
logfac.py.
