### 2020-01-15 3653506
Reduce number of test cases, add comment
Reduce the number of test cases and add a comment that the file was generated
by a script.
