+++ src/tests/test_Lambda.c
@@ -1,3 +1,4 @@
+/* file automatically generated by src/tests/test_Lambda.py */
 #include "libcaps.h"
 #include "unittest.h"
 
@@ -74,9 +75,7 @@ int test_caps_lnLambda(void)
     AssertAlmostEqual(&test, caps_lnLambda(50,50,2), -18.9156407761557381);
     AssertAlmostEqual(&test, caps_lnLambda(50,50,5), -42.4322399306116498);
     AssertAlmostEqual(&test, caps_lnLambda(50,50,10), -81.5362618302254081);
-    AssertAlmostEqual(&test, caps_lnLambda(50,50,20), -159.009535338258000);
     AssertAlmostEqual(&test, caps_lnLambda(50,50,50), -366.968103676874702);
-    AssertAlmostEqual(&test, caps_lnLambda(50,50,20), -159.009535338258000);
     AssertAlmostEqual(&test, caps_lnLambda(100,1,0), -1.75576034333105534);
     AssertAlmostEqual(&test, caps_lnLambda(100,1,1), -6.71247928502570340);
     AssertAlmostEqual(&test, caps_lnLambda(100,2,0), -2.04965367578211484);
@@ -99,61 +98,14 @@ int test_caps_lnLambda(void)
     AssertAlmostEqual(&test, caps_lnLambda(100,50,2), -20.6365049685871125);
     AssertAlmostEqual(&test, caps_lnLambda(100,50,5), -46.2233579889002297);
     AssertAlmostEqual(&test, caps_lnLambda(100,50,10), -88.8116922658618308);
-    AssertAlmostEqual(&test, caps_lnLambda(100,50,20), -173.532892414306376);
     AssertAlmostEqual(&test, caps_lnLambda(100,50,50), -413.713714184647814);
-    AssertAlmostEqual(&test, caps_lnLambda(100,50,20), -173.532892414306376);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,0), -3.91698579477027506);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,1), -13.1372764975996258);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,2), -22.3573691610184868);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,5), -50.0144760471888096);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,10), -96.0871227014982536);
-    AssertAlmostEqual(&test, caps_lnLambda(100,100,20), -188.056249490354752);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,50), -460.459324692420926);
     AssertAlmostEqual(&test, caps_lnLambda(100,100,100), -867.148972987175748);
-    AssertAlmostEqual(&test, caps_lnLambda(100,100,20), -188.056249490354752);
-    AssertAlmostEqual(&test, caps_lnLambda(200,1,0), -2.10109786959618943);
-    AssertAlmostEqual(&test, caps_lnLambda(200,1,1), -7.74848259717971829);
-    AssertAlmostEqual(&test, caps_lnLambda(200,2,0), -2.39499120204724893);
-    AssertAlmostEqual(&test, caps_lnLambda(200,2,1), -8.59168207396483264);
-    AssertAlmostEqual(&test, caps_lnLambda(200,2,2), -14.5856155155876265);
-    AssertAlmostEqual(&test, caps_lnLambda(200,3,0), -2.57332867401661512);
-    AssertAlmostEqual(&test, caps_lnLambda(200,3,1), -9.11659313621417149);
-    AssertAlmostEqual(&test, caps_lnLambda(200,3,2), -15.5686719437740429);
[…]
