### 2020-01-15 #2011 @eschnett
@michael-hartmann You might want to refer to a particular version number (release) of the software in the paper. The current release is 0.4.2, while the master branch is 0.5.
