### 2020-01-15 3ff17c6
Rename
Rename test_Lambda.{c,h} to test_lnLambda.{c,h}.
