### 2020-01-15 #2011 @michael-hartmann
@eschnett My plan is actually make the 0.5 release as soon as all the changes from the review are implemented.
