### 2020-01-15 9a783b7
Rename test_Plm.{c,h} to test_lnPlm.{c,h}
