+++ src/tests/test_lnLambda.c
@@ -2,7 +2,7 @@
 #include "libcaps.h"
 #include "unittest.h"
 
-#include "test_Lambda.h"
+#include "test_lnLambda.h"
 
 int test_caps_lnLambda(void)
 {
