### 2020-01-15 61cb20d
Add comment that file was partly generated by a script
