+++ src/tests/test_lnPlm.c
@@ -1,3 +1,10 @@
+/*
+ * Great parts of this file were automatically generated by
+ * src/python/test_plm.py. For especially high numbers of l and m, Mathematica
+ * was used to create more test cases for parameters where mpmath was not able
+ * to provide results.
+ */
+
 #include <math.h>
 #include "unittest.h"
 #include "plm.h"
