+++ src/tests/test_lnPlm.c
@@ -1,8 +1,8 @@
 /*
  * Great parts of this file were automatically generated by
- * src/python/test_plm.py. For especially high numbers of l and m, Mathematica
- * was used to create more test cases for parameters where mpmath was not able
- * to provide results.
+ * src/python/test_lnPlm.py. For especially high numbers of l and m,
+ * Mathematica was used to create more test cases for parameters where mpmath
+ * was not able to provide results.
  */
 
 #include "unittest.h"
