+++ src/tests/test_lnPlm.py
@@ -1,9 +1,4 @@
-#!/usr/bin/python
-
-from __future__ import division
 from mpmath import *
-
-mp.dps = 50
 
 def Plm(l,m,x):
     if m % 2 == 0:
@@ -13,19 +8,19 @@ def Plm(l,m,x):
 
 
 if __name__ == "__main__":
-    print("#include <math.h>")
+    mp.dps = 60
+
+    print("/* automatically generated by src/tests/test_plm.py */")
 
     print("#include \"unittest.h\"")
-    print("#include \"sfunc.h\"")
-    print("#include \"libcaps.h\"")
+    print("#include \"plm.h\"")
     print()
     print("#include \"test_Plm.h\"")
     print()
-    print("int test_Plm()")
+    print("int test_lnPlm()")
     print("{")
     print("    unittest_t test;")
-    print()
-    print("    unittest_init(&test, \"Plm\", \"Test associated Legendre polynomials\");")
+    print("    unittest_init(&test, \"lnPlm\", \"Test associated Legendre polynomials\", 1e-12);")
     print()
 
     for l in (1, 5, 10, 50, 100, 500, 1000, 2000, 5000, 10000, 15000, 20000, 25000):
@@ -34,9 +29,10 @@ if __name__ == "__main__":
                 if l >= m:
                     try:
                         value = Plm(l,m,mpf(x))
-                        print("    AssertAlmostEqual(&test, Plm(%d,%d,%g,1,1), %s);" % (l,m,x,float(log(abs(value)))))
+                        print("    AssertAlmostEqual(&test, lnPlm(%d,%d,%g), %s);" % (l,m,x,float(log(abs(value)))))
                     except:
-                        print("    // AssertAlmostEqual(&test, Plm(%d,%d,%g,1,1), ???);" % (l,m,x))
+                        #print("    // AssertAlmostEqual(&test, Plm(%d,%d,%g,1,1), ???);" % (l,m,x))
+                        pass
         print()
 
     print()
