### 2020-01-15 michael-hartmann/caps#157 @eschnett
Why are there lookup tables for `log` in the code?
I briefly looked at the code, and I find a large lookup table in `logfac.c` that seems to contain the values for `log(n)` for `n` from `0` to `65535`. Why are these tables there? Is this actually faster than calling the `log` function?

https://github.com/openjournals/joss-reviews/issues/2011
