### 2020-01-15 #2011 @eschnett
@michael-hartmann If you do, consider referencing this version in the paper. Also consider getting a DOI for that version (e.g. from Zenodo); I don't know whether JOSS recommends this anyway.
