### 2020-01-15 #2011 @eschnett
@michael-hartmann Do you describe the state of the art for numerical Casimir calculations in your paper? Are there other, similar packages available? What are their limitations?
