### 2020-01-15 michael-hartmann/caps#157 @michael-hartmann
The lookup tables are used for the functions `logi`, `lfac` and `lfac2` which are used in the files  `src/integration.c`, `src/libcaps.c`, and `src/plm.c.

Some time ago I was checking the speedup using a lookup table and found a significant performance gain for some special cases. I will check again if this is still the case.
