### 2020-01-15 6173f41
Remove not required header
