### 2020-01-15 721bc69
Add comment that reference data was obtained from Mathematica
