### 2020-01-15 3eded32
Adjust file to file renamings in the tests
