### 2020-01-15 #2011 @danielskatz
@eschnett  - part of the JOSS acceptance process is declaring a specific version (often new at that time), so don't worry about this too much at this point
