### 2020-01-15 #2011 @michael-hartmann
The first paragraph on the second page of the [proof](https://github.com/openjournals/joss-papers/blob/joss.02011/joss.02011/10.21105.joss.02011.pdf) describes the numerical problems that render the evaluation of the Casimir free energy difficult when the sphere radius is much larger than the separation between sphere and plate. (This case is experimentally the most interesting case.). The second paragraph on the second page describes how we solve these numerical issues.

I am not aware of any other available package to compute the Casimir free energy for this configuration. Also, nobody in the Casimir community has published exact numerical data for experimentally relevant parameters (there are lot of papers that use approximations, though.). Therefore, I think our approach is the state-of-the-art.

@eschnett Does this answer your question?
