### 2020-01-15 #2011 @eschnett
@michael-hartmann Do you want to add a note corresponding to your second paragraph above to the paper? This knowledge will tell others that your work is unique, instead of you not ignoring the competition.
