### 2020-01-18 michael-hartmann/caps#159 @michael-hartmann
Add remark to JOSS paper
Briefly mention in the JOSS paper that CaPS is currently the only openly available tool to compute the Casimir energy in the plane-sphere geometry.

openjournals/joss-reviews#2011
