### 2020-01-18 michael-hartmann/caps#159 @gertingold
There is also `MEEP` (https://meep.readthedocs.io/en/latest/) able to do Casimir calculations. Certainly, `MEEP` and `caps` have different domains of application. `MEEP` is more general purpose but not able to cover the aspect ratios handled by `caps`. The approach used by `MEEP` is also entirely different.
