### 2020-01-18 michael-hartmann/caps#157 @eschnett
Thanks for the detailed answer!

If you don't like the tables in the code, then you could calculate them at startup.
