### 2020-01-18 michael-hartmann/caps#159 @michael-hartmann
You are right. I didn‘t think of them because the focus is somewhat different. But I think we should  briefly mention them in the JOSS paper.
