### 2020-01-19 5cbee61
Add support for installing the programs
