### 2020-01-19 69bdd28
Add install option to README.md and user manual
