### 2020-01-19 michael-hartmann/caps#150 @michael-hartmann
I have added support for installing the programs unsing cmake. The README.md and the user manual have been adjusteed accordingly, see PR #160.

Ad the moment, only the executables and the library are installed. Before the header files can be installed, I have to clean them up first. I plan this for the next minor release, but for the JOSS review this requires too much work. Also, an uninstall option is missing as it seems that cmake does not support it out of the box.
