### 2020-01-19 124ba3a
Explain Casimir effect and purpose of CaPS
* An explanation of the Casimir has been added to the user manual.
* A very brief explanation of the Casimir has been added to README.md.

This commit addresses issue #153.
