+++ README.md
@@ -6,15 +6,18 @@
 
 ## What is CaPS?
 
-CaPS is an implementation of the **Ca**simir effect in the **p**lane-**s**phere
-geometry. The geometry consists of a sphere of radius R separated by a distance
-L from an infinite plate.
+CaPS is a package for the analysis of the **Ca**simir effect in the
+**p**lane-**s**phere geometry. The Casimir force arises due to quantum and
+thermal fluctuations of the electromagnetic field and is closely related to the
+van der Waals force.
 
-With CaPS you can compute the free Casimir energy in the plane-sphere geometry
-for arbitrary dielectrics at zero and finite temperature. The code is highly
-optimized and allows you - depending on parameters and your hardware - to
-compute the free energy for aspect ratios up to R/L~5'000 (in some cases even
-higher!).
+CaPS allows to compute the Casimir interaction in the plane-sphere geometry as
+a function of the sphere radius R, the minimal separation L between sphere and
+plane, the temperature T, and the material properties of plane and sphere. It
+is assumed that both objects are non-magnetic and placed in vacuum. The code is
+highly optimized and allows you - depending on parameters and the available
+resources - to compute the free energy for aspect ratios up to R/L~5'000 (in
+some cases even higher!).
 
 <img src="images/overview.svg" align="left" height="100%" width="100%" >
 
