### 2020-01-19 297994a
Slight changes
* Remove "you" as it is not necessary
* Write 5,000 instead of 5'000
