### 2020-01-26 #2011 @michael-hartmann
@jwuttke: Are you fine with the solution proposed for https://github.com/michael-hartmann/caps/issues/150?
