### 2020-01-26 903199f
Add configuration for readthedocs
