### 2020-01-26 581ba50
Revert "Add configuration for readthedocs"
This reverts commit 903199f3e293fe1920bbdd94c6c9d781bae8fc49.
