### 2020-01-26 9007063
Fix journal and authors for meep publication
