+++ docs/paper/paper.bib
@@ -267,9 +267,9 @@
 
 % meep
 @article{oskooi_cpc_2010,
-  author  = {Oskooi, A. and Roundy, D. and Ibanescu, M. and Bermel, P. and Joannopoulos, J. D. and Johnson, S. G.},
+  author  = {Oskooi, A. F. and Roundy, D. and Ibanescu, M. and Bermel, P. and Joannopoulos, J. D. and Johnson, S. G.},
   title   = {{Meep: A flexible free-software package for electromagnetic simulations by the FDTD method}},
-  journal = {Comput. Computer Physics Communications},
+  journal = {Comput. Phys. Commun.},
   volume  = {181},
   issue   = {3},
   pages   = {687--702},
