### 2020-01-26 30fd676
Rephrase paragraph
