### 2020-01-26 #2011 @michael-hartmann
@eschnett: We have added an additional paragraph mentioning two other packages that are able to compute the Casimir free energy and briefly describing the different focus. It is the second to last paragraph befor the acknowldgements starting with "Other packages that allow...".
