### 2020-01-26 #2011 @michael-hartmann
A big thank you to @eschnett and @jwuttke for taking time to review and your good suggestions. Also thank you @danielskatz for your work and time.

The necessary information:
 * The git tag is [v0.5](https://github.com/michael-hartmann/caps/releases/tag/v0.5).
 * zenodo DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3627969.svg)](https://doi.org/10.5281/zenodo.3627969)
