### 2020-01-26 #2011 @danielskatz
@michael-hartmann  Please merge https://github.com/michael-hartmann/caps/pull/163  Also, please check the cases of titles and journals in the references, using {}s to protect cases in the .bib file.  For example, in Buks, "casimer" in the title should probably be "Casimer". There's at least one other title with the same issue, and might be other titles that also need fixing.
