### 2020-01-26 #2011 @michael-hartmann
I just realized that the journal names are inconsistent: some are abbreviated, others not. I guess all journal names should be abbreviated?
