### 2020-01-26 b69d67c
Consistently abbreviate all journal names
