+++ docs/paper/paper.bib
@@ -74,7 +74,7 @@
 @article{bimonte_prl_2012,
   title   = {{Exact Results for Classical Casimir Interactions: Dirichlet and Drude Model in the Sphere-Sphere and Sphere-Plane Geometry}},
   author  = {Bimonte, Giuseppe and Emig, Thorsten},
-  journal = {Physical Review Letters},
+  journal = {Phys. Rev. Lett.},
   volume  = {109},
   issue   = {16},
   pages   = {160403},
@@ -90,7 +90,7 @@
 @article{lambrecht_njp_2006,
   title   = {{The Casimir effect within scattering theory}},
   author  = {Lambrecht, Astrid and Maia Neto, Paulo A. and Reynaud, Serge},
-  journal = {New Journal of Physics},
+  journal = {New J. Phys.},
   volume  = {8},
   number  = {10},
   pages   = {243},
@@ -102,7 +102,7 @@
 @article{emig_prl_2007,
   title   = {{Casimir Forces between Arbitrary Compact Objects}},
   author  = {Emig, T. and Graham, N. and Jaffe, R. L. and Kardar, M.},
-  journal = {Physical Review Letters},
+  journal = {Phys. Rev. Lett.},
   volume  = {99},
   issue   = {17},
   pages   = {170403},
@@ -119,7 +119,7 @@
 @article{ambikasaran_josc_2013,
   title     = {{An $\mathcal O (N \log N)$ Fast Direct Solver for Partial Hierarchically Semi-Separable Matrices}},
   author    = {Ambikasaran, Sivaram and Darve, Eric},
-  journal   = {Journal of Scientific Computing},
+  journal   = {J. Sci. Comput.},
   volume    = {57},
   number    = {3},
   pages     = {477--501},
@@ -147,7 +147,7 @@
   volume    = {4},
   number    = {34},
   pages     = {1167},
-  journal   = {The Journal of Open Source Software},
+  journal   = {J. Open Source Softw.},
   doi       = {10.21105/joss.01167}
 }
 
@@ -159,7 +159,7 @@
 @article{ingold_pre_2015,
   title   = {{Geometric origin of negative Casimir entropies: A scattering-channel analysis}},
   author  = {Ingold, Gert-Ludwig and Umrath, Stefan and Hartmann, Michael and Gu\'erout, Romain and Lambrecht, Astrid and Reynaud, Serge and Milton, Kimball A.},
-  journal = {Physical Review E},
+  journal = {Phys. Rev. E},
   volume  = {91},
   issue   = {3},
   pages   = {033203},
@@ -171,7 +171,7 @@
 @article{umrath_pre_2015,
   title   = {{Disentangling geometric and dissipative origins of negative Casimir entropies}},
   author  = {Umrath, Stefan and Hartmann, Michael and Ingold, Gert-Ludwig and Maia Neto, Paulo A.},
-  journal = {Physical Review E},
+  journal = {Phys. Rev. E},
   volume  = {92},
   issue   = {4},
   pages   = {042125},
@@ -183,7 +183,7 @@
 @article{hartmann_prl_2017,
   title   = {{Plasma versus Drude Modeling of the Casimir Force: Beyond the Proximity Force Approximation}},
   author  = {Hartmann, Michael and Ingold, Gert-Ludwig and Maia Neto, Paulo A.},
-  journal = {Physical Review Letters},
+  journal = {Phys. Rev. Lett.},
   volume  = {119},
   issue   = {4},
   pages   = {043901},
@@ -204,7 +204,7 @@
 @article{hartmann_phscr_2018,
   author  = {Hartmann, Michael and Ingold, Gert-Ludwig and {Maia Neto}, Paulo A.},
[…]
