### 2020-01-26 #2011 @michael-hartmann
Okay, then let's go for all journal names abbreviated.

@gertingold Can you have a look at the PR https://github.com/michael-hartmann/caps/pull/165 and merge it if it's okay?
