### 2020-01-26 #2011 @gertingold
Is there an established style for citing JOSS? Should it be "JOSS" or would "J. Open Source Softw." also be ok?
