### 2020-01-26 #2011 @gertingold
@danielskatz, we have just merged a commit consistently abbreviating all journal titles. The only open question concerns the abbreviation of JOSS (see my previous commit). If "J. Open Source Softw." is fine, all journal titles should now be consistent. If JOSS is preferred, we will make one more change.
