### 2020-01-26 6bd0716
Abbreviate Journal of Open Source Software as JOSS
