+++ docs/paper/paper.bib
@@ -147,7 +147,7 @@
   volume    = {4},
   number    = {34},
   pages     = {1167},
-  journal   = {J. Open Source Softw.},
+  journal   = {JOSS},
   doi       = {10.21105/joss.01167}
 }
 
