### 2020-01-26 7e1df2c
made journal titles unabbreviated
