+++ docs/paper/paper.bib
@@ -7,7 +7,7 @@
   title   = {{Short-range fundamental forces}},
   author  = {I. Antoniadis and S. Baessler and M. B\"uchner and V.V. Fedorov and S. Hoedl and A. Lambrecht and V.V. Nesvizhevsky and G. Pignol and K.V. Protasov and S. Reynaud and Yu. Sobolev},
   author  = {Antoniadis, I. and Baessler, S. and B\"uchner, M. and Fedorov, V. V. and Hoedl, S. and Lambrecht, A. and Nesvizhevsky, V. V. and Pignol, G. and Protasov, K. V. and Reynaud, S. and Sobolev, Yu.},
-  journal = {C. R. Phys.},
+  journal = {Comptes Rendus Physique},
   volume  = {12},
   number  = {8},
   pages   = {755--778},
@@ -19,7 +19,7 @@
 @article{chan_prl_2001,
   title   = {{Nonlinear Micromechanical Casimir Oscillator}},
   author  = {Chan, H. B. and Aksyuk, V. A. and Kleiman, R. N. and Bishop, D. J. and Capasso, Federico},
-  journal = {Phys. Rev. Lett.},
+  journal = {Physical Review Letters},
   volume  = {87},
   issue   = {21},
   pages   = {211801},
@@ -31,7 +31,7 @@
 @article{buks_prb_2001,
   title   = {{Stiction, adhesion energy, and the Casimir effect in micromechanical systems}},
   author  = {Buks, E. and Roukes, M. L.},
-  journal = {Phys. Rev. B},
+  journal = {Physical Review B},
   volume  = {63},
   issue   = {3},
   pages   = {033402},
@@ -53,7 +53,7 @@
 @article{martin_crp_2012,
   title   = {{Everything you always wanted to know about the cosmological constant problem (but were afraid to ask)}},
   author  = {Martin, J.},
-  journal = {C. R. Phys.},
+  journal = {Comptes Rendus Physique},
   volume  = {13},
   number  = {6},
   pages   = {566--665},
@@ -74,7 +74,7 @@
 @article{bimonte_prl_2012,
   title   = {{Exact Results for Classical Casimir Interactions: Dirichlet and Drude Model in the Sphere-Sphere and Sphere-Plane Geometry}},
   author  = {Bimonte, Giuseppe and Emig, Thorsten},
-  journal = {Phys. Rev. Lett.},
+  journal = {Physical Review Letters},
   volume  = {109},
   issue   = {16},
   pages   = {160403},
@@ -90,7 +90,7 @@
 @article{lambrecht_njp_2006,
   title   = {{The Casimir effect within scattering theory}},
   author  = {Lambrecht, Astrid and Maia Neto, Paulo A. and Reynaud, Serge},
-  journal = {New J. Phys.},
+  journal = {New Journal of Physics},
   volume  = {8},
   number  = {10},
   pages   = {243},
@@ -102,7 +102,7 @@
 @article{emig_prl_2007,
   title   = {{Casimir Forces between Arbitrary Compact Objects}},
   author  = {Emig, T. and Graham, N. and Jaffe, R. L. and Kardar, M.},
-  journal = {Phys. Rev. Lett.},
+  journal = {Physical Review Letters},
   volume  = {99},
   issue   = {17},
   pages   = {170403},
@@ -119,7 +119,7 @@
 @article{ambikasaran_josc_2013,
   title     = {{An $\mathcal O (N \log N)$ Fast Direct Solver for Partial Hierarchically Semi-Separable Matrices}},
   author    = {Ambikasaran, Sivaram and Darve, Eric},
-  journal   = {J. Sci. Comput.},
+  journal   = {Journal of Scientific Computing},
   volume    = {57},
   number    = {3},
[…]
