### 2020-01-26 #2011 @gertingold
We just moved to unabbreviated titles to prevent this issue... In any case, journal titles are now consistent.
